Reorganize hintable impl
