+++ ceno_recursion/src/tower_verifier/binding.rs
@@ -125,102 +125,6 @@ impl Hintable<InnerConfig> for PointAndEval {
 }
 impl VecAutoHintable for PointAndEval {}
 
-#[derive(Debug, Deserialize)]
-pub struct IOPProverMessage {
-    pub evaluations: Vec<E>,
-}
-
-use sumcheck::structs::IOPProverMessage as InnerIOPProverMessage;
-impl From<InnerIOPProverMessage<E>> for IOPProverMessage {
-    fn from(value: InnerIOPProverMessage<E>) -> Self {
-        IOPProverMessage {
-            evaluations: value.evaluations,
-        }
-    }
-}
-
-impl Hintable<InnerConfig> for IOPProverMessage {
-    type HintVariable = IOPProverMessageVariable<InnerConfig>;
-
-    fn read(builder: &mut Builder<InnerConfig>) -> Self::HintVariable {
-        IOPProverMessageVariable {
-            evaluations: builder.hint_exts(),
-        }
-    }
-
-    fn write(&self) -> Vec<Vec<<InnerConfig as Config>::N>> {
-        let mut stream = Vec::new();
-        stream.extend(self.evaluations.write());
-        stream
-    }
-}
-impl VecAutoHintable for IOPProverMessage {}
-
-/// Assume that all the prover messages have the same size.
-#[derive(Debug, Deserialize, Default)]
-pub struct IOPProverMessageVec {
-    pub prover_message_size: usize,
-    pub data: Vec<E>,
-}
-
-impl IOPProverMessageVec {
-    pub fn get(&self, index: usize) -> &[E] {
-        let start = index * self.prover_message_size;
-        let end = start + self.prover_message_size;
-        &self.data[start..end]
-    }
-}
-
-impl From<Vec<IOPProverMessage>> for IOPProverMessageVec {
-    fn from(messages: Vec<IOPProverMessage>) -> Self {
-        let prover_message_size = messages[0].evaluations.len();
-        assert!(messages
-            .iter()
-            .map(|message| message.evaluations.len())
-            .all_equal());
-        let data = messages
-            .into_iter()
-            .flat_map(|msg| msg.evaluations)
-            .collect();
-        IOPProverMessageVec {
-            prover_message_size,
-            data,
-        }
-    }
-}
-
-impl Hintable<InnerConfig> for IOPProverMessageVec {
-    type HintVariable = IOPProverMessageVecVariable<InnerConfig>;
-
-    fn read(builder: &mut Builder<InnerConfig>) -> Self::HintVariable {
-        let prover_message_size: Var<F> = usize::read(builder);
-        let length: Var<F> = usize::read(builder);
-        let evaluations = Vec::<E>::read(builder);
-        builder.assert_eq::<Var<F>>(evaluations.len(), prover_message_size * length);
-        IOPProverMessageVecVariable {
-            prover_message_size,
-            length,
-            evaluations,
-        }
-    }
-
-    fn write(&self) -> Vec<Vec<<InnerConfig as Config>::N>> {
-        let mut stream = Vec::new();
-        stream.extend(<usize as Hintable<InnerConfig>>::write(
-            &self.prover_message_size,
-        ));
-        stream.extend(<usize as Hintable<InnerConfig>>::write(
-            &if self.data.len() == 0 {
-                0
-            } else {
-                self.data.len() / self.prover_message_size
-            },
-        ));
-        stream.extend(self.data.write());
-        stream
-    }
-}
-
 #[derive(Debug, Default, Deserialize)]
 pub struct ThreeDimensionalVector {
     pub inner_inner_length: usize,
